pyproject.toml
[tool.poetry]
name = "fastapi-starter"
version = "0.1.0"
description = ""
authors = ["Miguel Angel Zetina Zetina <miguel.zetina@storicard.com>"]
readme = "README.md"
packages = [{include = "fastapi_starter"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
fastapi-utils = "^0.2.1"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
pydantic = {extras= ["dotenv"], version = "^1.10.2"}
dataclasses-json = "^0.5.7"
gpl = "^0.1.4"
strawberry-graphql = {extras = ["debug-server"], version = "^0.151.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
